Issue #4587: Moved all resources for TrailingComment check
