+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/TrailingCommentCheckTest.java
@@ -45,7 +45,10 @@ public class TrailingCommentCheckTest extends BaseCheckTestSupport {
 
     @Override
     protected String getPath(String filename) throws IOException {
-        return super.getPath("checks" + File.separator + filename);
+        return super.getPath("checks" + File.separator
+                + "misc" + File.separator
+                + "trailingcomment" + File.separator
+                + filename);
     }
 
     @Test
